review 1.2

+++ backend/api/views.py
@@ -11,9 +11,8 @@ from rest_framework.response import Response
 from .filters import IngredientFilter, RecipeFilter
 from users.models import User, Subscription
 from .permissions import AuthorOrReadOnly
-from recipes.models import (Favorite, Ingredient, Recipe,
-                            RecipeIngredient,
-                            ShoppingCart, Tag)
+from recipes.models import (Ingredient, Recipe,
+                            RecipeIngredient, Tag)
 from .serializers import (CreateRecipeSerializer, UserSerializer,
                           FavoriteSerializer, IngredientSerializer,
                           RecipeSerializer, ShoppingCartSerializer,
@@ -45,7 +44,7 @@ class UserViewSet(UserViewSet):
 
     @action(
         detail=True,
-        methods=['post', 'delete'],
+        methods=['post'],
         url_path='subscribe',
         url_name='subscribe',
     )
@@ -63,9 +62,16 @@ class UserViewSet(UserViewSet):
         if not Subscription.objects.filter(subscriber=request.user,
                                            author=author).exists():
             raise ValidationError('Такой подписки нет')
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+    @to_subscribe.mapping.delete
+    def delete_subscription(self, request, id):
+        author = get_object_or_404(User, id=id)
+        if not Subscription.objects.filter(subscriber=request.user,
+                                           author=author).exists():
+            raise ValidationError('Такой подписки нет')
         Subscription.objects.filter(subscriber=request.user,
                                     author=author).delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(
         detail=False,
@@ -105,22 +111,6 @@ class RecipeViewSet(viewsets.ModelViewSet):
     filterset_class = RecipeFilter
 
     def get_queryset(self):
-        user = self.request.user
-        if user.is_authenticated:
-            return (Recipe.objects.all().select_related('author')
-                    .annotate(is_favorited=Exists(
-                        Favorite.objects.filter(recipe=OuterRef('pk'))
-                        .select_related('user')
-                    ))
-                    .annotate(is_in_shopping_cart=Exists(ShoppingCart.objects
-                                                         .filter(
-                                                             recipe=OuterRef(
-                                                                 'pk'))
-                                                         .select_related(
-                                                             'user')
-                                                         ))
-                    .prefetch_related('tags', 'ingredients')
-                    )
         return (Recipe.objects.all().select_related('author')
                 .prefetch_related('tags', 'ingredients'))
 
@@ -131,7 +121,7 @@ class RecipeViewSet(viewsets.ModelViewSet):
 
     @action(
         detail=True,
-        methods=['post', 'delete'],
+        methods=['post'],
         url_path='shopping_cart',
         url_name='shopping_cart',
     )
@@ -150,14 +140,24 @@ class RecipeViewSet(viewsets.ModelViewSet):
                 recipe=get_object_or_404(Recipe, pk=pk))
             if not queryset.exists():
                 return Response(status=status.HTTP_400_BAD_REQUEST)
-            queryset.delete()
-        else:
+        elif not self.request.user.is_authenticated:
             return Response(status=status.HTTP_401_UNAUTHORIZED)
         return Response(status=status.HTTP_204_NO_CONTENT)
 
+    @get_shopping_cart.mapping.delete
+    def delete_shopping_cart(self, request, pk):
+        if self.request.user.is_authenticated:
+            queryset = ShoppingCartSerializer.Meta.model.objects.filter(
+                user=request.user,
+                recipe=get_object_or_404(Recipe, pk=pk))
+            if not queryset.exists():
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+            queryset.delete()
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+
     @action(
         detail=True,
-        methods=['post', 'delete'],
+        methods=['post'],
         url_path='favorite',
         url_name='favorite',
     )
@@ -176,10 +176,18 @@ class RecipeViewSet(viewsets.ModelViewSet):
                 recipe=get_object_or_404(Recipe, pk=pk))
             if not queryset.exists():
                 return Response(status=status.HTTP_400_BAD_REQUEST)
-            queryset.delete()
         else:
             return Response(status=status.HTTP_401_UNAUTHORIZED)
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+    @get_favorite.mapping.delete
+    def delete_favorite(self, request, pk):
+        queryset = FavoriteSerializer.Meta.model.objects.filter(
+            user=request.user,
+            recipe=get_object_or_404(Recipe, pk=pk))
+        if not queryset.exists():
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        queryset.delete()
 
     @action(
         detail=False,
